pyproject.toml
[tool.poetry]
name = "qpystructs"
version = "0.1.0"
description = "fluent structs is a tools to access,transfer different data type"
authors = ["fluentqa <KUZz8@example.com>"]
packages =[
    {from="src", include="qpystructs"}
]
[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
jmespath = "^1.0.1"
deepdiff = "^6.3.0"
pyexcel = "^0.7.0"
openpyxl = "^3.1.2"
dotty-dict = "^1.3.1"
pyexcel-xlsx = "^0.6.0"
inflection = "^0.5.1"
PyYAML = "^6.0"
pandas = "^2.0.2"
pydantic-yaml = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
bump-pydantic = "^0.7.0"
coverage-badge = "^1.1.0"
isort = "^5.12.0"
black = "^23.11.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# https://github.com/psf/black
target-version = ["py310","py311","py312"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 310
line_length = 88
known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true
force_single_line = true
combine_as_imports = true
lines_between_types = 1
lines_after_imports = 2
src_paths = ["src", "tests"]
extend_skip = ["setup.py"]
